+++ debug/handler.go
@@ -15,8 +15,3 @@ type Handler interface {
 	// HandleFunc registers the handler function for the given pattern.
 	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
 }
-
-// NewHandler creates a new HTTP debug handler.
-func NewHandler() Handler {
-	return http.NewServeMux()
-}